Take Damage Logic

+++ Assets/sgu/Script/Player/PlayerState.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 public class PlayerState : MonoBehaviour
 {
     [SerializeField]
-    private int health = 100;
-    public int Health { get { return health; } set { health = value; } }
+    private float health = 100;
+    public float Health { get { return health; } set { health = value; } }
 
 }
